requirements.txt
cloudpickle==3.0.0
easyocr==1.7.1
Farama-Notifications==0.0.4
filelock==3.15.4
fsspec==2024.6.1
gymnasium==0.29.1
imageio==2.34.2
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==2.0.0
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.3.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
PyYAML==6.0.1
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.4
six==1.16.0
sympy==1.12.1
tifffile==2024.6.18
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.2
